Refactor item class, adding base mod classes
Working out the best way to implement some of the modding classes, more
to come in future, and I'm working out the best way to handle this
stuff.

I am aware some classes are empty, this is mostly a structural thing to
pave the way for working on mods in general.