RozWorld icon now appears in the window bar
...using Windows specific code!

Yeah it's not great, but as of right now, it's the only thing I see
working, and it *does* work.

Also, there's a new GameTime timer and Update() function. These will not
only be used to track window focus, most likely the keyboard and mouse
state stuff will be updated to use this function instead of the current
system.