Fix font loading, game builds again. other updates
Font loading works again, there's some changes with Files.cs for texture
pack stuff.

The game builds again, the last commit broke it in a weird way, this
should fix it.

Also updated a bunch of the comments to have the right oddmatics domain
(it changed from .co.uk to .uk a while ago).