Changing language loading to directory based
The langs.ini file in the /link folder should now point to directories
that contain language dictionaries.

This is for better mod support, instead of a user having to edit the
language directories themselves, mods will be able to have language
.inis packed with them that the user just drag-drops into the right
folder. Or if the author decides to recreate the folder structure with
empty folders, the user just merges everything.

Basically it's nice and it will load all .ini files from the destination
folder. :neckbeard: