Removing COMFY and BasicObjects, new texture system
This update will remove COMFY and the BasicObjects folders. This move is
because of the creation of the API will include all the necessary mod
classes and the client should only store load routines and clientside
handling of these objects.

As a result of COMFY's removal, RozWorld will now load textures via .ini
files placed in the /link folder as long as the file starts with the
prefix "texmap_". This is so that mods that introduce new objects and
thus need to reference new textures can simply replicate the folder
structure and include a /link/texmap_*.ini that gets merged when the
user drags the files into their game directory.

The .ini files read as dictionaries in the code, in order to add
textures just follow the guide of:
TextureName:ImagePath

For example:
Splash:\gui\logos\OddSplash.png

In the code, GetTexture("Splash") would retrieve the texture generated
from the OddSplash.png file.

Also note that the ImagePath should start in the root of a texturepack,
do not include a texturepack's folder in the path (eg. the above starts
at "\gui" and not "\roz\gui").