Removing old files, FONTS KIND OF IMPLEMENTED
Good news!

Fonts now work, they draw to a bitmap perfectly... I think. There's just
a few little touch ups to do (mostly to GUIOMETRY.BIN since there's now
a thing to test the fonts against).

Right now there's some commented out code that when uncommented will
output some test results to a .png in the C:\ drive. This was just to
see what the drawing routine came up with, it appears to be perfect as
far as I can see, so the next move is to apply some changes (the default
colour should be white, not black) and make ZetaLabel use it).