Ported to GLFW from FreeGLUT - full project sync
Sorry it's been ages since the last update, it's taken me a while to
feel back on track but most of today has been ditching FreeGLUT because
it sucks BAD.

GLFW is the new way to go, it's crunch time! Note though, a lot of stuff
is broken right now, the game will build but input will be getting
tweaked heavily for instance. Some render changes will be applied now
and then I will crunch the API and client and all that. I am working
towards a project deadline so I would really love to get as much of it
done ASAP.

Basically, I'm a lazy bugger and it's coming back to bite me in the arse
(check my commit history for proof).